+++ modules/apps/object/object-rest-api/bnd.bnd
@@ -8,6 +8,6 @@ Export-Package:\
 	com.liferay.object.rest.filter.parser,\
 	com.liferay.object.rest.manager.exception,\
 	com.liferay.object.rest.manager.v1_0,\
-	com.liferay.object.rest.odata.entity.v1_0,\
+	com.liferay.object.rest.odata.entity.v1_0.provider,\
 	com.liferay.object.rest.openapi.v1_0,\
 	com.liferay.object.rest.resource.v1_0
