+++ modules/apps/object/object-rest-api/src/main/java/com/liferay/object/rest/odata/entity/v1_0/provider/EntityModelProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.rest.odata.entity.v1_0;
+package com.liferay.object.rest.odata.entity.v1_0.provider;
 
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.portal.odata.entity.EntityModel;
