+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/odata/entity/v1_0/ObjectEntryEntityModel.java
@@ -49,7 +49,7 @@ public class ObjectEntryEntityModel implements EntityModel {
 		return _entityFieldsMap;
 	}
 
-	protected ObjectEntryEntityModel(
+	public ObjectEntryEntityModel(
 			ObjectDefinition objectDefinition, List<ObjectField> objectFields)
 		throws Exception {
